I worked on a lot of stuff because I was free today.  I can do more tomorrow as well, but the main ones I didn't get to were outputting files as strings, which isn't hard, just tedious, and creating a gradebook with a given string (that's a little trickier).
I know the initializewithfile is pretty long, but I spent most of the
day on it and wanted to make progress else where.  I can fix it up
tomorrow though.  Everything else should be pretty good though in terms
of length.

Another problem I had was with the initializewithfile; in one file she
uses integers as inputs, but in another, she uses doubles.  I tried to
account for that, but my idea didn't really work (convertoint), so
hopefully you guys will have an idea about that.